refactor: isolate application error taxonomy
